04-cloud/requirements.txt
functions-framework
google-cloud-bigquery
google-cloud-storage
PyYAML
pandas
pyarrow